pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abbfn2"
version = "0.1.0"
description = "Official Repo of AbBFN2."
authors = [
    { name = "InstaDeep", email = "m.braganca@instadeep.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiobotocore==2.17.0",
    "boto3",
    "biopython==1.84",
    "distrax==0.1.4",
    "jax==0.4.34",
    "jaxlib==0.4.34",
    "flax==0.8.4",
    "hydra-core",
    "numpy",
    "omegaconf",
    "orbax-checkpoint==0.5.20",
    "s3fs==2024.6.1",
    "tabulate",
    "tqdm",
    "h5py==3.11.0",
    "setuptools",
    "regex",
    "cloudpathlib",
    "humatch @git+https://github.com/oxpig/Humatch.git"
]
